Move all test helper types and funcs into single file
